refactor: Split application/job/pipeline.rs (2.8k lines) into a pipelin… [ORB-12270]
Split application/job/pipeline.rs (2.8k lines) into a pipeline/ module by concern

Planned-By: codex
Orchestrated-By: astra
